# driver-location/docs/swagger.yaml
definitions:
...
  models.DriverLocation:
    properties:
      distance:
        type: number
      id:
        type: string
      location:
        $ref: '#/definitions/models.GeoJSON'
    type: object
...
  models.GeoJSON:
    properties:
      coordinates:
        items:
          type: number
        type: array
      type:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /driver/nearest:
    post:
      consumes:
      - application/json
      description: GetNearestDriver  and returns the driver location and distance
        of the driver
      parameters:
      - description: models.GeoJSON credentials
        in: body
        name: userLocation
        required: true
        schema:
          $ref: '#/definitions/models.GeoJSON'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DriverLocation'
        "400":
          description: Bad Request
          schema: {}
      summary: GetNearestDriver Endpoint
      tags:
      - GetNearestDriver
...
swagger: "2.0"